% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_algorithm0.R
\name{allocation_national}
\alias{allocation_national}
\title{allocation_national}
\usage{
allocation_national(recip_matrix, donor_matrix, HLA)
}
\arguments{
\item{recip_matrix}{a data.frame with recipient features, this is the input recipient matrix}

\item{donor_matrix}{a data.frame with donor features,this is the input donor matrix}

\item{HLA}{a matrix, this is the calculated HLA mismatching counts matrix}
}
\value{
a matrix contains allo_score, which is the alocation score for each new pair
}
\description{
this is the Australia current national allocation algorithm
}
\examples{
data("rawdata", package = "simKAP")


allocation_score <- allocation_national(rawdata,rawdata,rawdata);
}
