% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_algorithm0.R
\name{australia_state_selection}
\alias{australia_state_selection}
\title{Australia_state_selectionpool}
\usage{
australia_state_selection(
  recip_matrix,
  donor_matrix,
  whichdonor,
  bloodGroup_strict_match = TRUE
)
}
\arguments{
\item{recip_matrix}{a data.frame with recipient features, this is the input recipient matrix}

\item{donor_matrix}{a data.frame with recipient features, this is the input donor matrix}

\item{whichdonor}{a categorical donor id, this is the incoming kidney from one donor}

\item{bloodGroup_strict_match}{a boolean value, TRUE/FALSE, this is to define whether we want the strict match for bloodgroup or not}
}
\value{
a data.frame representing recipient pool under the state allocation tule
}
\description{
Australia_state_selectionpool
}
\examples{
data("rawdata", package = "simKAP")
state_pool <- australia_state_selection(rawdata,rawdata,"D948826",
bloodGroup_strict_match = TRUE);
}
