% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blg_match.R
\name{blood_group_match}
\alias{blood_group_match}
\title{blood group match}
\usage{
blood_group_match(
  donor_bl,
  recipient_bl,
  strict_match = TRUE,
  AB_priority = TRUE
)
}
\arguments{
\item{donor_bl}{a character value,this is the donor blood group}

\item{recipient_bl}{a character value,this is the irecipient blood group}

\item{strict_match}{a boolean value, TRUE/FALSE, this is the strict match indicator}

\item{AB_priority}{a boolean value, TRUE/FALSE, this is the AB priority indicator}
}
\value{
a data.frame with matched recipient and donor
}
\description{
this is the blood group match function
}
\examples{
data("rawdata", package = "simKAP")


blgmatch <- blood_group_match(rawdata$donor_blgroup,rawdata$recip_blgroup,TRUE,TRUE);
}
