% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection0.R
\name{dm_national_formula}
\alias{dm_national_formula}
\title{selection}
\usage{
dm_national_formula(
  recip_matrix = NULL,
  donor_matrix = NULL,
  allocation_score = NULL,
  HLA = NULL,
  graft_number = 1,
  parameter_a = 1,
  parameter_b = 1,
  parameter_c = 1,
  quality = TRUE,
  age = TRUE,
  state = TRUE,
  bloodgroup = TRUE,
  accept_prob = NULL,
  yearly_scale = rep(1, 100)
)
}
\arguments{
\item{recip_matrix}{a data.frame with recipient features, this is the input recipient matrix}

\item{donor_matrix}{a data.frame with donor features, this is the input donor matrix}

\item{allocation_score}{a vector, this is the score for each pair}

\item{HLA}{a matrix, this is the calculated HLA mismatching counts matrix}

\item{graft_number}{a numerical value, this is the graft number}

\item{parameter_a}{an interger hyperparameter to tune}

\item{parameter_b}{an interger hyperparameter to tune}

\item{parameter_c}{an interger hyperparameter to tune}

\item{quality}{a boolean value, TRUE/FALSE, quality measure indicator}

\item{age}{a boolean value, TRUE/FALSE, age indicator}

\item{state}{a boolean value, TRUE/FALSE, state indicator}

\item{bloodgroup}{a boolean value, TRUE/FALSE, blood group indicator}

\item{accept_prob}{a numerical value between 0 and 1, which is a user defined minimum acceptence probability, then for different recipient, we randomly sample from this value to 1, the default value is NULL, which means we give an accept rate based on HLA matchability}

\item{yearly_scale}{a vector representing yearly scale, this allows the user to define different accpetence rate for different years}
}
\value{
a data.frame with matched recipient
}
\description{
this is the sdm1
}
\examples{
data("newdata", package = "KidneyAllocation")
HLA_matrix <- hla_match(raw_recip_matrix_subset, raw_donor_matrix[1,])

selected <- dm_national_formula(raw_recip_matrix_subset,raw_donor_matrix[1,],
allocation_score=allocation_national(raw_recip_matrix_subset,raw_donor_matrix[1,],HLA_matrix),
HLA_matrix,1,1,1,1,TRUE,TRUE,TRUE,TRUE,accept_prob = NULL, yearly_scale=rep(1, 100));
}
