% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection0.R
\name{dm_state_top}
\alias{dm_state_top}
\title{selection}
\usage{
dm_state_top(
  recip_matrix = NULL,
  donor_matrix = NULL,
  allocation_score = NULL,
  HLA = NULL,
  graft_number = 1,
  parameter_state_a = 1,
  parameter_state_b = 1,
  bloodgroup = 1.5,
  age = 1.5,
  top_state_matches = 15,
  parameter_state = 10
)
}
\arguments{
\item{recip_matrix}{a data.frame with recipient features, this is the input recipient matrix}

\item{donor_matrix}{a data.frame with donor features, this is the input donor matrix}

\item{allocation_score}{a vector, this is the score for each pair}

\item{HLA}{a matrix, this is the calculated HLA mismatching counts matrix}

\item{graft_number}{a numerical value, this is the graft number}

\item{parameter_state_a}{an integer hyperparameter to tune}

\item{parameter_state_b}{an integer hyperparameter to tune}

\item{bloodgroup}{a boolean value, TRUE/FALSE, blood group indicator}

\item{age}{a boolean value, TRUE/FALSE, age indicator}

\item{top_state_matches}{an integer, top match parameter}

\item{parameter_state}{an integer, state parameter}
}
\value{
a data.frame with matched recipient
}
\description{
this is the sdm2
}
\examples{
data("newdata", package = "simKAP")
HLA_matrix <- hla_match(raw_recip_matrix_subset, raw_donor_matrix[1,])

selected <- dm_state_top(raw_recip_matrix_subset,raw_donor_matrix[1,],
allocation_score=allocation_national(raw_recip_matrix_subset,raw_donor_matrix[1,],HLA_matrix),
HLA_matrix,1,1,1.5,1.5,15,10);

}
