% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl0.R
\name{dynamic_waitlist}
\alias{dynamic_waitlist}
\title{wl function}
\usage{
dynamic_waitlist(
  donor_matrix,
  yearly_scale_in = NULL,
  yearly_scale_out = NULL,
  rate_in = 0.7,
  rate_out = 0.4,
  IN = FALSE,
  OUT = FALSE,
  waitlist_Risk = TRUE
)
}
\arguments{
\item{donor_matrix}{a data.frame, this is the input donor matrix}

\item{yearly_scale_in}{a numeirc vector between 0 and 1, this is the input yearly scale rate}

\item{yearly_scale_out}{a numeirc vector between 0 and 1, this is the input yearly scale rate}

\item{rate_in}{a numeirc value between 0 and 1, this is the rate}

\item{rate_out}{a numeirc value between 0 and 1, this is the rate}

\item{IN}{a boolean value, TRUE/FALSE, whether to use the user defined rate_in or not}

\item{OUT}{a boolean value, TRUE/FALSE, whether to use the user defined rate_out or not}

\item{waitlist_Risk}{a boolean value, TRUE/FALSE, whether to have this or not}
}
\value{
wl data which is a matrix for all recipients on the waiting list
}
\description{
this is the dynamic wl function
}
\examples{
data("rawdata", package = "KidneyAllocation")


wl <- dynamic_waitlist(rawdata,yearly_scale_in = NULL,yearly_scale_out = NULL,rate_in = 0.7,
IN= TRUE, rate_out = 0.4,waitlist_Risk = TRUE);

}
