% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{hla_match}
\alias{hla_match}
\title{HLA functions}
\usage{
hla_match(Rmatrix, Dvector)
}
\arguments{
\item{Rmatrix}{a matrix, this is the recipient matrix}

\item{Dvector}{a matrix, this is the donor matrix}
}
\value{
HLA mismatching matrix
}
\description{
this is the function for calculating HLA
}
\examples{
data("rawdata", package = "simKAP")

HLA_matrix <- hla_match(rawdata, rawdata);

}
