% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linMap.R
\name{linMap}
\alias{linMap}
\title{proportion mapping function}
\usage{
linMap(x, from, to)
}
\arguments{
\item{x}{this is the input score vector}

\item{from}{a numerical value, this is the lower bound}

\item{to}{a numerical value, this is the upper bound}
}
\value{
another vector with percentage ranking score
}
\description{
this is changing the score to percentage ranking
}
\examples{
data("rawdata", package = "KidneyAllocation")


new_score <- linMap(rawdata$recip_epts,1,100);
}
