% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_donor_matrix}
\alias{raw_donor_matrix}
\title{Example data for this package}
\format{
An object of class \code{data.frame} with 2835 rows and 28 columns.
}
\source{
ANZDATA.
}
\usage{
data(newdata, package = 'simKAP')
}
\description{
Example data for this package
}
\keyword{datasets}
