% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_recip_matrix_subset}
\alias{raw_recip_matrix_subset}
\title{Example data for this package}
\format{
An object of class \code{data.frame} with 1477 rows and 49 columns.
}
\source{
ANZDATA.
}
\usage{
data(newdata, package = 'simKAP')
}
\description{
Example data for this package
}
\keyword{datasets}
