% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Example data for this package}
\format{
An object of class \code{data.frame} with 5945 rows and 80 columns.
}
\source{
ANZDATA.
}
\usage{
data(rawdata, package = 'simKAP')
}
\description{
Example data for this package
}
\keyword{datasets}
