% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection0.R
\name{select_max}
\alias{select_max}
\title{selection}
\usage{
select_max(
  recip_matrix = NULL,
  donor_matrix = NULL,
  allocation_score,
  HLA = NULL,
  graft_number = 1
)
}
\arguments{
\item{recip_matrix}{a data.frame with recipient features, this is the input recipient matrix}

\item{donor_matrix}{a data.frame with donor features, this is the input donor matrix}

\item{allocation_score}{a vector, this is the score for each pair}

\item{HLA}{a matrix, this is the calculated HLA mismatching counts matrix}

\item{graft_number}{a numerical value, this is the graft number}
}
\value{
a matrix contains allo_score, which is the alocation score for each new pair
}
\description{
this is the max selection function
}
\examples{
data("rawdata", package = "simKAP")
selected <- select_max(allocation_score=allocation_national(rawdata,rawdata,rawdata),
graft_number = 1);
}
