% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool0.R
\name{selection_default}
\alias{selection_default}
\title{national_selection_pool}
\usage{
selection_default(
  recip_matrix,
  donor_matrix,
  whichdonor,
  bloodGroup_strict_match = TRUE,
  AB_priority = TRUE
)
}
\arguments{
\item{recip_matrix}{a data.frame, this is the input recipient matrix}

\item{donor_matrix}{a data.frame, this is the input donor matrix}

\item{whichdonor}{a categorical donor id, this is the incoming kidney from one donor}

\item{bloodGroup_strict_match}{a boolean value, TRUE/FALSE, this is to define whether we want the strict match for bloodgroup or not}

\item{AB_priority}{a boolean value, TRUE/FALSE, this is to define whether AB group can get from other blood group or not}
}
\value{
a data.frame contains eligible recipients
}
\description{
this is the Australia current national recipient selection pool
}
\examples{

data("rawdata", package = "KidneyAllocation")
rawdata$donor_rank = rank(rawdata[,'donor_kdri']) / nrow(rawdata)


selectionpool1.2 <- selection_default(rawdata,rawdata,1,
bloodGroup_strict_match = TRUE,AB_priority = TRUE);

}
