% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool0.R
\name{selection_irisk}
\alias{selection_irisk}
\title{au_slicingwindow_selection_pool (updated recip epts for the new tx_date)}
\usage{
selection_irisk(
  recip_matrix,
  donor_matrix,
  threshold_number = 0.1,
  whichdonor,
  bloodGroup_strict_match = TRUE,
  AB_priority = TRUE
)
}
\arguments{
\item{recip_matrix}{a data.frame, this is the input recipient matrix}

\item{donor_matrix}{a data.frame, this is the input donor matrix}

\item{threshold_number}{a numerical value between 0 and 1, this is the threshold number we would like to use for matching}

\item{whichdonor}{a categorical donor id, this is the incoming kidney from one donor}

\item{bloodGroup_strict_match}{a boolean value, TRUE/FALSE, this is to define whether we want the strict match for bloodgroup or not}

\item{AB_priority}{a boolean value, TRUE/FALSE, this is to define whether AB group can get from other blood group or not}
}
\value{
a data.frame contains eligible recipients
}
\description{
this is the au with threshold matching selection pool
}
\examples{
data("rawdata", package = "KidneyAllocation")
rawdata$donor_rank = rank(rawdata[,'donor_kdri']) / nrow(rawdata)

rawdata$recip_tx_date=rawdata$tx_date
selectionpool1.1 <- selection_irisk(rawdata,rawdata,threshold_number = 0.1,
whichdonor=1,bloodGroup_strict_match = TRUE,AB_priority = TRUE);
}
